Fix form persist for IE11/Edge
