--- skins/laika/src/view_models/DataPersistenceItem.ts
@@ -1,11 +1,0 @@
-export enum DataPersistenceMutationType {
-	VALUE,
-	KEY_VALUE_PAIR,
-}
-
-export interface DataPersistenceItem {
-	key: string,
-	mutationType: DataPersistenceMutationType,
-	mutation: string,
-	fields: string[]
-}
