+++ skins/laika/tests/unit/DataPersister.spec.ts
@@ -1,4 +1,4 @@
-import { DataPersister } from '@/store/data_persistence/DataPersister';
+import { ActiveDataPersister as DataPersister } from '@/store/data_persistence/ActiveDataPersister';
 import FakeDataPersistenceRepository from './TestDoubles/FakeDataPersistenceRepository';
 import { FakeDataEncryptor, FakeFailingDataEncryptor } from './TestDoubles/FakeDataEncryptor';
